move get npmrc path
