--- test/unit/io/npmrc-io.test.ts
@@ -1,16 +1,0 @@
-import path from "path";
-import { getHomeNpmrcPath } from "../../../src/io/npmrc-io";
-
-describe("npmrc-io", () => {
-  describe("find path in home", () => {
-    const someHomePath = path.join(path.sep, "user", "dir");
-
-    it("should be [Home]/.npmrc", () => {
-      const expected = path.join(someHomePath, ".npmrc");
-
-      const actual = getHomeNpmrcPath(someHomePath);
-
-      expect(actual).toEqual(expected);
-    });
-  });
-});
